Avancements

+++ ProjetIntegrateur/src/logiqueCircuit/Resistance.java
@@ -8,12 +8,12 @@ public class Resistance implements Composante{
     final int ID;
     double ohm;
     
-    public Resistance(int ohm, int id) {
+    public Resistance(double ohm, int id) {
         ID = GestionnaireID.getInstance().ajouterComp(id);
         this.ohm = ohm;
     }
 
-    public Resistance(int ohm) {
+    public Resistance(double ohm) {
         ID = GestionnaireID.getInstance().ajouterComp();
         this.ohm = ohm;
     }
